# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Simple Books REST api",
    "version": "1.0.0",
    "title": "Swagger Books ",
    "contact": {
      "email": "andrej.delmar@gmail.com"
    }
  },
  "host": "localhost",
  "basePath": "/",
  "tags": [
    {
      "name": "Book",
      "description": "Book controller"
    },
    {
      "name": "Author",
      "description": "Author controller"
    }
  ],
  "schemes": ["https", "http"],
  "paths": {
    "/books": {
      "post": {
        "tags": ["Book"],
        "summary": "Create new book with an author",
        "operationId": "createBook",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Book with author",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "409": {
            "description": "Conflict"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": ["Book"],
        "summary": "Get books by filter",
        "operationId": "getBooks",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "title", "type": "string" },
          { "in": "query", "name": "description", "type": "string" },
          { "in": "query", "name": "isbn", "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/books/{bookId}": {
      "patch": {
        "tags": ["Book"],
        "summary": "Update a book with an author",
        "operationId": "updateBook",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "deleteAuthors",
            "type": "string",
            "description": "true = deletes omitted authors from this book"
          },
          {
            "in": "path",
            "name": "bookId",
            "description": "ID of the targeted book",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Book with author",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": ["Book"],
        "summary": "Delete a book",
        "operationId": "updateBook",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "description": "ID of the targeted book",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": ["Book"],
        "summary": "Get a book",
        "operationId": "updateBook",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "description": "ID of the targeted book",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "tags": ["Author"],
        "summary": "Get authors by filter",
        "operationId": "updateBook",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "firstName", "type": "string" },
          { "in": "query", "name": "lastName", "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "tags": ["Author"],
        "summary": "Get author with books",
        "operationId": "getAuthor",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the targeted author",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": ["Author"],
        "summary": "Delete an author",
        "operationId": "deleteAuthor",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the targeted author",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Night Watch"
        },
        "description": {
          "type": "string",
          "example": "For a policeman, there can be few things worse than a serial killer at loose in your city. Except, perhaps, a serial killer who targets coppers, and a city on the brink of bloody revolution. The people have found their voice at last, the flags and barricades are rising...And the question for a policeman, an officer of the law, a defender of the peace, is: Are you with them, or are you against them?"
        },
        "isbn": {
          "type": "string",
          "example": "9780552154307"
        },
        "authors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Author"
          }
        }
      }
    },
    "Author": {
      "type": "object",
      "required": ["firstName", "lastName"],
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Terry"
        },
        "lastName": {
          "type": "string",
          "example": "Pratchett"
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}